fixes for track_rs=1 and interframe/thread statefullness
-fixed startup values of dx9 renderstates
-fixed SR/BF on iframe start and replay thread transfer
-fixed separatealphablend state
-added SetSamplerState tracked version

+++ d912pxy/d912pxy_batch.cpp
@@ -206,7 +206,7 @@ void d912pxy_batch::GPUWriteControlMT(UINT64 si, UINT64 of, UINT64 cnt, UINT64 b
 
 		ctl->dstOffset = (UINT32)i;
 		ctl->startBatch = (UINT32)bn;
-		ctl->endBatch = endDlt;
+		ctl->endBatch = (UINT32)endDlt;
 
 		--endDlt;
 		++si;
